refactor(CartesianChart): update import paths for type definitions
